refactor: Phase 2 조직·구조 정돈 (D1, D2, D3, C3)
- D1: src/view/ 신설, 컬러/메시지 포맷터 통합
  (colors.js, testMessages.js, errorMessages.js)
- D2: constants/index.js를 paths.js·module.js로 분리,
  NUM 상수 제거(리터럴 0/1 직접 사용)
- D3: src/utils/, src/mock/utils/ 디렉터리 제거
- C3: mock/utils/changeModuleExports.js를
  makeMockFnc.js·changeModuleExports.js로 분할
  (mockFunctions 헬퍼는 makeMockFnc.js 내부로 흡수)
