+++ index.js
@@ -1,7 +1,7 @@
 import {testManager} from "./src/testManager.js";
 import {clearAllMocks, isMocked, mock, unmock, mockStore} from './src/mock/store.js';
 import {expect} from "./src/expect.js";
-import {makeMockFnc} from "./src/mock/utils/changeModuleExports.js";
+import {makeMockFnc} from "./src/mock/makeMockFnc.js";
 
 /**
  * 테스트 케이스를 정의합니다.
