+++ src/cli/utils/findFiles.js
@@ -1,8 +1,6 @@
 import fs from 'fs';
 import path from 'path';
-
-
-import {PATH} from "../../constants/index.js";
+import {PATH} from "../../constants/paths.js";
 
 /**
  * 테스트 파일을 찾는 로직입니다.
