+++ src/constants/module.js
@@ -1,0 +1,5 @@
+export const MODULE_TYPE = {
+  MODULE: 'module',
+  ESM: 'esm',
+  CJS: 'cjs',
+};
