+++ src/view/colors.js
@@ -1,4 +1,12 @@
-import {COLORS} from "../constants/view.js";
+const COLORS = {
+  reset: '\x1b[0m',
+  green: '\x1b[32m',
+  red: '\x1b[31m',
+  yellow: '\x1b[33m',
+  cyan: '\x1b[36m',
+  gray: '\x1b[90m',
+  bold: '\x1b[1m',
+};
 
 export const colorize = (text, color) => `${color}${text}${COLORS.reset}`;
